+++ examples/nextjs-security-demo/pages/_error.tsx
@@ -1,0 +1,5 @@
+export default function ErrorPage() {
+  return <p>Something went wrong</p>;
+}
+
+